pyproject.toml
[project]
name = "drive-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.171.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "platformdirs>=4.3.8",
    "rapidfuzz>=3.13.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.403",
    "ruff>=0.12.5",
]

[tool.pyright]
typeCheckingMode = "standard"
exclude = [".venv/", "__pycache__/"]
venv = ".venv"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
